Bug fixing and feature additions

See merge request RSE-Lab-Fasilkom-UI/PricesIDE/payment-gateway/vmj-payment-gateway!6

+++ modules/paymentgateway.disbursement.core/paymentgateway/disbursement/core/service/DisbursementServiceImpl.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import java.util.*;
 import java.util.logging.Logger;
 import java.io.File;
+import java.io.IOException;
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
@@ -18,6 +19,7 @@ import vmj.routing.route.exceptions.*;
 import paymentgateway.disbursement.DisbursementFactory;
 import paymentgateway.config.core.Config;
 import paymentgateway.config.ConfigFactory;
+import org.hibernate.hql.internal.ast.QuerySyntaxException;
 
 public class DisbursementServiceImpl extends DisbursementServiceComponent {
 	private static final Logger LOGGER = Logger.getLogger(DisbursementServiceImpl.class.getName());
@@ -26,7 +28,7 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 		String workingDir = System.getProperty("user.dir");
 		List<File> propertyFiles = new ArrayList<>();
 		List<String> vendors = new ArrayList<>();
-		String[] targetFiles = {"oy.properties", "flip.properties", "midtrans.properties"};
+		String[] targetFiles = {"oy.properties", "flip.properties", "midtrans.properties", "xendit.properties"};
 
 		// Iterate through target files
 		for (String targetFile : targetFiles) {
@@ -45,30 +47,33 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 	            Map<String, Object> requestMap = config.getCallbackDisbursementRequestBody(requestBody);
 
 	            String idStr = (String) requestMap.get("id");
+				String vendorGeneratedIdStr = (String) requestMap.get("vendor_generated_id");
 	            String status = (String) requestMap.get("status");
 
 	            LOGGER.info("Processing Vendor: " + vendor);
 	            LOGGER.info("ID: " + idStr);
+				LOGGER.info("Vendor Generated ID: " + vendorGeneratedIdStr);
 	            LOGGER.info("Status: " + status);
 
-				String hostAddress = getEnvVariableHostAddress("AMANAH_HOST_BE");
-        		int portNum = getEnvVariablePortNumber("AMANAH_PORT_BE");
-	            HttpClient client = HttpClient.newHttpClient();
-				String configUrl = String.format("http://%s:%d/call/receivedisbursementcallback", hostAddress, portNum);
-	            String requestString = config.getRequestString(requestMap);
-	            HttpRequest request = config.getBuilder(HttpRequest.newBuilder(), config.getHeaderParams())
-	                                       .uri(URI.create(configUrl))
-	                                       .POST(HttpRequest.BodyPublishers.ofString(requestString))
-	                                       .build();
-				
-	            try {
-	                HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-	                String rawResponse = response.body();
-	                LOGGER.info("Raw Response: " + rawResponse);
-	            } catch (Exception e) {
-	                System.err.println("Failed to send request for vendor: " + vendor);
-	                e.printStackTrace();
-	            }
+				Disbursement disbursement = null;
+				if (idStr != null) {
+					disbursement = this.getObject(idStr);
+				} else {
+					List<Disbursement> disbursements = Repository.getAllObject("disbursement_impl");
+					for (Disbursement d : disbursements) {
+						if (d.getVendorGeneratedId().equals(vendorGeneratedIdStr)) {
+							disbursement = d;
+						}
+					}
+				}
+
+				if (disbursement == null) {
+					throw new BadRequestException("Disbursement record not found");
+				}
+
+				disbursement.setStatus(status.toUpperCase());
+		
+				this.updateObject(disbursement);
 	        } catch (Exception e) {
 	            System.err.println("Failed to process vendor: " + vendor);
 	            e.printStackTrace();
@@ -79,27 +84,34 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 	}
 
     public Disbursement createDisbursement(Map<String, Object> requestBody) {
-        Map<String, Object> response = sendTransaction(validateRequestBody(requestBody));
+		this.validateVendorName((String) requestBody.get("vendor_name"));
+		this.validateRequiredStringField(requestBody, "bank_code");
+		this.validateRequiredStringField(requestBody, "account_number");
+		this.validateAmount(requestBody.get("amount"));
+        Map<String, Object> response = sendTransaction(requestBody);
         return createDisbursement(requestBody, response);
     }
 	
 	public Disbursement createDisbursement(Map<String, Object> requestBody, Map<String, Object> response){
-		Map<String, Object> validatedRequestBody = validateRequestBody(requestBody);
-		String bank_code = (String) validatedRequestBody.get("bank_code");
-		String account_number = (String) validatedRequestBody.get("account_number");
-		double amount = (Double) validatedRequestBody.get("amount");
-		int id = (int) response.get("id");
+		String vendorName = this.validateVendorName((String) requestBody.get("vendor_name"));
+		String bank_code = this.validateRequiredStringField(requestBody, "bank_code");
+		String account_number = this.validateRequiredStringField(requestBody, "account_number");
+		double amount = this.validateAmount(requestBody.get("amount"));
+		String id = (String) response.get("id");
 		int userId = (int) response.get("user_id");
 		String status = (String) response.get("status");
+		String vendorGeneratedId = (String) response.get("vendor_generated_id");
 		
 		Disbursement disbursement = DisbursementFactory.createDisbursement(
 			"paymentgateway.disbursement.core.DisbursementImpl",
-			id,
+			UUID.fromString(id),
 			userId,
 			account_number,
 			amount,
 			bank_code,
-			status
+			status.toUpperCase(),
+			vendorName,
+			vendorGeneratedId
 		);
 
 		Repository.saveObject(disbursement);
@@ -108,17 +120,19 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 	}
 	
 	public HashMap<String, Object> updateDisbursement(Map<String, Object> requestBody) {
-
-		int id = ((Double) requestBody.get("id")).intValue();
+		String id = this.validateId(requestBody.get("id"));
 		Disbursement disbursement = this.getObject(id);
 
-		try {
-			disbursement.setAmount((Double) requestBody.get("amount"));
-			disbursement.setAccountNumber((String) requestBody.get("account_number"));
-			disbursement.setBankCode((String) requestBody.get("bank_code"));
-		} catch (Exception e){
-			e.printStackTrace();
+		if (disbursement == null) {
+			throw new BadRequestException(String.format("Disbursement with ID %s does not exist", id));
 		}
+
+		double amount = this.validateAmount(requestBody.get("amount"));
+		String accountNumber = this.validateRequiredStringField(requestBody, "account_number");
+		String bankCode = this.validateRequiredStringField(requestBody, "bank_code");
+		disbursement.setAmount(amount);
+		disbursement.setAccountNumber(accountNumber);
+		disbursement.setBankCode(bankCode);
 
 		this.updateObject(disbursement);
 		
@@ -127,11 +141,40 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
     }
 	
 	public List<HashMap<String, Object>> deleteDisbursement(Map<String, Object> requestBody){
-		int id = ((Double) requestBody.get("id")).intValue();
+		String id = this.validateId(requestBody.get("id"));
 		Disbursement disbursement = this.getObject(id);
-		this.deleteObject(id);
+		
+		if (disbursement == null) {
+			throw new BadRequestException("Disbursement dengan ID " + id + " tidak ditemukan");
+		}
 
-		return getAllDisbursement(requestBody);
+		final String[] disbursementHolder = {null};
+		Repository.executeQuery(session -> {
+			String sql = String.format("SELECT modulesequence FROM disbursement_comp WHERE id ='%s'", id );
+			try {
+				String result = (String) session.createNativeQuery(sql).getSingleResult();
+				String[] modules = result.split(",");
+				disbursementHolder[0] = modules[modules.length - 1].trim();
+			} catch (Exception e) {
+				disbursementHolder[0] = "";
+			}
+		});
+
+		if (disbursementHolder[0].equals("disbursement_impl")) {
+			this.deleteObject(id);
+			return getAllDisbursement(disbursementHolder[0]);
+		}
+
+		final String[] targetId = {""};
+		Repository.executeQuery(session -> {
+			String sql = String.format("SELECT cast(id as varchar) FROM %s WHERE base_component_id = '%s'", disbursementHolder[0], id);
+			targetId[0] = (String) session.createNativeQuery(sql).getSingleResult();
+		});
+
+		
+		this.deleteObject(targetId[0]);
+
+		return getAllDisbursement(disbursementHolder[0]);
 	}
 	
 	public String getEnvVariableHostAddress(String varname_host){
@@ -146,7 +189,8 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 	}
 
 	public Map<String, Object> sendTransaction(Map<String, Object> requestBody) {
-        String vendorName = (String) requestBody.get("vendor_name");
+		String id = UUID.randomUUID().toString();
+        String vendorName = this.validateVendorName((String) requestBody.get("vendor_name"));
 		Config config = ConfigFactory.createConfig(vendorName,
 				ConfigFactory.createConfig("paymentgateway.config.core.ConfigImpl"));
 
@@ -168,18 +212,18 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 			HttpResponse response = client.send(request, HttpResponse.BodyHandlers.ofString());
 			String rawResponse = response.body().toString();
 			LOGGER.info("Raw Response: " + rawResponse);
-			responseMap = config.getDisbursementResponse(rawResponse);
-		} catch (Exception e) {
+			responseMap = config.getDisbursementResponse(rawResponse, id);
+		} catch (IOException | InterruptedException e) {
 			e.printStackTrace();
 		}
 
 		return responseMap;
 	}
 	
-	public HashMap<String, Object> getDisbursementById(int id){
+	public HashMap<String, Object> getDisbursementById(String id){
 		List<HashMap<String, Object>> disbursementList = getAllDisbursement("disbursement_impl");
 		for (HashMap<String, Object> disbursement : disbursementList){
-			int record_id = ((Double) disbursement.get("record_id")).intValue();
+			String record_id = (String) disbursement.get("record_id");
 			if (record_id == id){
 				return disbursement;
 			}
@@ -188,22 +232,40 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 		return null;
 	}
 
-	public List<HashMap<String, Object>> getAllDisbursement(String tableName){
-		List<Disbursement> List = Repository.getAllObject(tableName);
-		return transformListToHashMap(List);
+	public HashMap<String, Object> findById(List<HashMap<String, Object>> disbursements, String id) {
+		for (HashMap<String, Object> disbursement : disbursements){
+			String disbursementId = (String) disbursement.get("id");
+			if (disbursementId.equals(id)){
+				return disbursement;
+			}
+		}
+
+		throw new BadRequestException("Disbursement dengan ID " + id + " tidak ditemukan");
+	}
+
+	public List<HashMap<String, Object>> getAllDisbursement(){
+		return this.getAllDisbursement("disbursement_impl");
 	}
 	
-	public HashMap<String, Object> getDisbursement(Map<String, Object> requestBody){
-		int id = ((Double) requestBody.get("id")).intValue();
+	public HashMap<String, Object> getDisbursement(String id) {
+		this.validateId(id);
 		Disbursement disbursementImpl = this.getObject(id);
-		HashMap<String, Object> disbursementDataMap = disbursementImpl.toHashMap();
-		return disbursementDataMap;
+
+	    if (disbursementImpl == null) {
+	    	throw new BadRequestException("Disbursement dengan ID " + id + " tidak ditemukan");
+	    }
+	    
+		return disbursementImpl.toHashMap();
 	}
 	
-	public List<HashMap<String, Object>> getAllDisbursement(Map<String, Object> requestBody){
-		String table = (String) requestBody.get("table_name");
-		List<Disbursement> List = Repository.getAllObject(table);
-		return transformListToHashMap(List);
+	public List<HashMap<String, Object>> getAllDisbursement(String tableName){
+		
+		try {
+			List<Disbursement> list = Repository.getAllObject(tableName);
+		    return transformListToHashMap(list);
+		} catch (Exception e) {
+			throw new BadRequestException("Table name " + tableName + " bukan entity yang valid");
+		}
 	}
 
 	public String getParamsUrlEncoded(Map<String, Object> requestBody){
@@ -237,12 +299,12 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
 	
 	
 	
-	public Disbursement getObject(int id) {
-        return Repository.getObject(id);
+	public Disbursement getObject(String id) {
+        return Repository.getObject(UUID.fromString(id));
     }
 
-    public void deleteObject(int id) {
-        Repository.deleteObject(id);
+    public void deleteObject(String id) {
+        Repository.deleteObject(UUID.fromString(id));
     }
 
     public void updateObject(Disbursement disbursement) {
@@ -254,9 +316,83 @@ public class DisbursementServiceImpl extends DisbursementServiceComponent {
     }
 
 	private Map<String, Object> validateRequestBody(Map<String, Object> requestBody) {
-		String vendorName = (String) requestBody.get("vendor_name");
+		String vendorName = validateVendorName((String) requestBody.get("vendor_name"));
 		Config config = ConfigFactory.createConfig(vendorName,
 				ConfigFactory.createConfig("paymentgateway.config.core.ConfigImpl"));
         return config.getDisbursementRequestBody(requestBody);
 	}
+
+	public String validateVendorName(String vendorName) {
+		if (vendorName == null) {
+			throw new BadRequestException("vendor_name tidak ditemukan pada payload");
+		}
+		try {
+			// Provide vendor name that supports dibursement only
+			System.out.println("Vendor Name:" + vendorName);
+			Set<String> vendorNames = new HashSet<>();
+			vendorNames.add("Flip");
+			vendorNames.add("Xendit");
+			
+			if (!vendorNames.contains(vendorName)) {
+				throw new BadRequestException("vendor_name tidak valid.");
+			}
+			return vendorName;
+		} catch (Exception e) {
+			throw new BadRequestException("vendor_name tidak valid");
+		}
+
+	}
+
+	public double validateAmount(Object amountObject) {
+		Double amount;
+		if (amountObject == null) {
+			throw new BadRequestException("amount tidak ditemukan pada payload.");
+		}
+		try {
+			amount = ((Double) amountObject);
+		} catch (Exception e) {
+			try {
+				String amountString = (String) amountObject;
+				amount = Double.valueOf(amountString);
+			} catch (Exception ex) {
+				throw new BadRequestException("amount tidak valid.");
+			}
+		}
+
+		if (amount < 0) {
+			throw new BadRequestException("amount tidak boleh negatif.");
+		}
+
+		return amount.doubleValue();
+	}
+	
+	public String validateId(Object idObject) {
+		String id;
+		if (idObject == null) {
+			throw new BadRequestException("id tidak ditemukan pada payload.");
+		}
+		try {
+			id = ((String) idObject);
+			UUID.fromString(id);
+			return id;
+		} catch (Exception e) {
+			throw new BadRequestException("id tidak valid.");
+		}
+	}
+	
+	public String validateRequiredStringField(Map<String, Object> requestBody, String key) {
+		Object field = requestBody.get(key);
+		if (field == null) {
+			throw new BadRequestException(String.format("%s tidak ditemukan pada payload.", key));
+		}
+
+		
+		String stringField = (String) field;
+		if (stringField.length() == 0) {
+			throw new BadRequestException(String.format("%s tidak boleh berupa string kosong.", key));
+		} 
+
+		return stringField;
+	}
+
 }
